Add a JUnit 4 test runner
This is not strictly necessary but will ease integration with existing
tools (IDE, CI...).

Please notice that the tests all pass and that they have a very high
coverage ratio: 100% lines on the GildedRose class. The Item class
appears to have no coverage at all but it is not accurate: the
toString() method is indeed never called because we register another
formatter which will give a more precise output. The constructor *IS*
invoked though, but maybe in a way that is not compatible with IntelliJ's
integrated coverage tool.