Move conditional to upper level
Although this changes the execution path slightly, it does not impact
our test results because the observable state will still be the same
after our change.

A key element in knowing this change is safe is knowing that the
conditions do not have any side effect.

The only effect the change will have on the code is that the processor
will have to process a few more conditionals than before... Until a
profiling tool proves otherwise it is *NOT* a problem.

Remember Donald Knuth: "premature optimization is the root of all evil".