Get rid of a magic number discrepancy
While the requirements mention that something happens to backstage
passes 10 and 5 days before the event, the code uses constants 11 and 6.

Although the link seems pretty obvious I think it is a better practice
to have code constants that reflect the requirements, so we introduce a
little change that will help us do it.

Note: once again we *did* change the execution flow, but did *not*
change the observable state. The tests still pass.