Extract some methods
The code is much more readable this way.
It still has some duplication, but let's take a few minutes to compare
it to its initial state.

Now wouldn't be a bad time to add the new "Conjured Item" feature, but
we will do a little more refactoring before considering it.