Move methods
GildedRose should not be responsible for the implementation of the Item
operations. Because we can't modify the Item class, we introduce a
helper class: ItemOperations.

Note: the helper class is in fact an enum: although it is Serializable,
no instances can be created outside of the enum and no class can extend
it. Both characteristics are suitable for helper classes.